Final for validation

+++ src/main/java/tn/insat/jebouquine/presentation/controller/administrator/AuteurController.java
@@ -1,6 +1,7 @@
 package tn.insat.jebouquine.presentation.controller.administrator;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -8,19 +9,24 @@ import org.springframework.web.servlet.ModelAndView;
 import tn.insat.jebouquine.business.facade.IGestionAuteur;
 import tn.insat.jebouquine.data.entity.Auteur;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 /**
  * Created by Devcartha on 12/14/2015.
  */
 @Controller
+@Scope(value = "session")
 public class AuteurController {
     @Autowired
     IGestionAuteur gestionAuteur;
 
     @RequestMapping(value = "/auteur/formI", method = RequestMethod.GET)
     @ResponseBody
-    public ModelAndView getAuteurFormForInsertion() {
+    public ModelAndView getAuteurFormForInsertion(HttpSession httpSession) {
+        if(httpSession.getAttribute("admin")==null){
+            return new ModelAndView("redirect:/admin/loginForm");
+        }
         ModelAndView mv = new ModelAndView("administratorView/auteur/form");
         mv.addObject("auteur", new Auteur());
         return mv;
@@ -28,14 +34,20 @@ public class AuteurController {
 
     @RequestMapping(value = "/auteur/formU", method = RequestMethod.GET)
     @ResponseBody
-    public ModelAndView getAuteurFormForUpdate(@RequestParam Long id) {
+    public ModelAndView getAuteurFormForUpdate(@RequestParam Long id, HttpSession httpSession) {
+        if(httpSession.getAttribute("admin")==null){
+            return new ModelAndView("redirect:/admin/loginForm");
+        }
         ModelAndView mv = new ModelAndView("administratorView/auteur/form");
         mv.addObject("auteur", gestionAuteur.getAuteurById(id));
         return mv;
     }
 
     @RequestMapping(value = "/auteur/save", method = RequestMethod.POST)
-    public ModelAndView addAuteur(@ModelAttribute @Valid Auteur auteur, BindingResult bindingResult) {
+    public ModelAndView addAuteur(@ModelAttribute @Valid Auteur auteur, BindingResult bindingResult, HttpSession httpSession) {
+        if(httpSession.getAttribute("admin")==null){
+            return new ModelAndView("redirect:/admin/loginForm");
+        }
         if (bindingResult.hasErrors()) {
             return new ModelAndView("administratorView/auteur/form");
         } else {
@@ -45,7 +57,10 @@ public class AuteurController {
     }
 
     @RequestMapping(value = "/auteur/update", method = RequestMethod.POST)
-    public ModelAndView updateAuteur(@ModelAttribute @Valid Auteur auteur, BindingResult bindingResult) {
+    public ModelAndView updateAuteur(@ModelAttribute @Valid Auteur auteur, BindingResult bindingResult, HttpSession httpSession) {
+        if(httpSession.getAttribute("admin")==null){
+            return new ModelAndView("redirect:/admin/loginForm");
+        }
         if (bindingResult.hasErrors()) {
             return new ModelAndView("administratorView/auteur/form");
         } else {
@@ -57,7 +72,10 @@ public class AuteurController {
 
     @RequestMapping(value = "/auteur/search", method = RequestMethod.GET)
     @ResponseBody
-    public ModelAndView searchAuteur(@RequestParam String keyWord) {
+    public ModelAndView searchAuteur(@RequestParam String keyWord, HttpSession httpSession) {
+        if(httpSession.getAttribute("admin")==null){
+            return new ModelAndView("redirect:/admin/loginForm");
+        }
         ModelAndView mv = new ModelAndView("administratorView/auteur/list");
         mv.addObject("auteurs",gestionAuteur.getAuteursByKeyWord(keyWord));
         return mv;
@@ -66,7 +84,10 @@ public class AuteurController {
 
     @RequestMapping(value = "/auteur/list", method = RequestMethod.GET)
     @ResponseBody
-    public ModelAndView getAuteursList() {
+    public ModelAndView getAuteursList(HttpSession httpSession) {
+        if(httpSession.getAttribute("admin")==null){
+            return new ModelAndView("redirect:/admin/loginForm");
+        }
         ModelAndView mv = new ModelAndView("administratorView/auteur/list");
         mv.addObject("auteurs", gestionAuteur.getAll());
         return mv;
@@ -75,7 +96,10 @@ public class AuteurController {
 
     @RequestMapping(value = "/auteur/delete", method = RequestMethod.GET)
     @ResponseBody
-    public ModelAndView deleteAuteur(@RequestParam Long id) {
+    public ModelAndView deleteAuteur(@RequestParam Long id, HttpSession httpSession) {
+        if(httpSession.getAttribute("admin")==null){
+            return new ModelAndView("redirect:/admin/loginForm");
+        }
         gestionAuteur.deleteAuteur(id);
         return new ModelAndView("redirect:/auteur/list");
     }
